[IMP] fieldservice_stock: Split stock request
