+++ fieldservice_stock/models/__init__.py
@@ -5,7 +5,6 @@ from . import (
     fsm_territory,
     fsm_location,
     fsm_order,
-    stock_move_line,
     stock_move,
     procurement_group,
     stock_rule,
@@ -15,6 +14,4 @@ from . import (
     product_template,
     stock_production_lot,
     fsm_wizard,
-    stock_request,
-    stock_request_order,
 )
