+++ fieldservice_stock/views/stock.xml
@@ -1,12 +1,6 @@
 <odoo>
 
     <!-- Menu Items -->
-    <menuitem id="menu_fsm_stock_request"
-              name="Stock Requests"
-              action="stock_request.action_stock_request_form"
-              parent="fieldservice.operations"
-              sequence="90"/>
-
     <menuitem id="menu_fsm_product"
               name="Products"
               action="stock.product_template_action_product"
