+++ fieldservice_stock_request/models/__init__.py
@@ -1,0 +1,6 @@
+from . import (
+    fsm_order,
+    stock_move_line,
+    stock_request,
+    stock_request_order,
+)
