+++ fieldservice_stock_request/models/stock_move_line.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2018 Brian McMaster
+# Copyright (C) 2018 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from odoo import exceptions, models
