+++ fieldservice_stock_request/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Brian McMaster <brian@mcmpest.com>
+* Sandip Mangukiya <smangukiya@opensourceintegrators.com>
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
