+++ fieldservice_stock_request/readme/CREDITS.rst
@@ -1,0 +1,3 @@
+The development of this module has been financially supported by:
+
+* Open Source Integrators <https://opensourceintegrators.com>
