+++ fieldservice_stock_request/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module is an add-on for the Field Service application in Odoo.
+It provides stock requests in field service orders.
