+++ fieldservice_stock_request/readme/ROADMAP.rst
@@ -1,0 +1,2 @@
+The roadmap of the Field Service application is documented on
+`Github <https://github.com/OCA/field-service/issues/1>`_.
