+++ fieldservice_stock_request/views/stock.xml
@@ -1,0 +1,10 @@
+<odoo>
+
+    <!-- Menu Items -->
+    <menuitem id="menu_fsm_stock_request"
+              name="Stock Requests"
+              action="stock_request.action_stock_request_form"
+              parent="fieldservice.operations"
+              sequence="90"/>
+
+</odoo>
