+++ fieldservice_stock_request/views/stock_request.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <record id="stock_request_fsm_order_form" model="ir.ui.view">
